Removing autoloot functions for client 10x

+++ src/item.cpp
@@ -583,8 +583,8 @@ Attr_ReadValue Item::readAttr(AttrTypes_t attr, PropStream& propStream)
 		}
 
 		case ATTR_OPENCONTAINER: {
-			int32_t openContainer;
-			if (!propStream.read<int32_t>(openContainer)) {
+			uint8_t openContainer;
+			if (!propStream.read<uint8_t>(openContainer)) {
 				return ATTR_READ_ERROR;
 			}
 
@@ -839,7 +839,7 @@ void Item::serializeAttr(PropWriteStream& propWriteStream) const
 
 	if (hasAttribute(ITEM_ATTRIBUTE_OPENCONTAINER)) {
 		propWriteStream.write<uint8_t>(ATTR_OPENCONTAINER);
-		propWriteStream.write<int32_t>(getIntAttr(ITEM_ATTRIBUTE_OPENCONTAINER));
+		propWriteStream.write<uint8_t>(getIntAttr(ITEM_ATTRIBUTE_OPENCONTAINER));
 	}
 
 	if (hasAttribute(ITEM_ATTRIBUTE_ARMOR)) {
